+++ internal/expvar/doc.go
@@ -1,7 +1,4 @@
 // Package expvar is the standard Golang's expvar package with one modification:
-// it removes the bad practice of using `init()`, instead exposes:
-// - `Start`: does everything the old `init` does
-// - `PublishCmdLine`: publishes command line information
-// - `PublishMemStats`: publishes memory statistics
-// - `RegisterHandler`: registers the standard endpoint: `GET /debug/vars`.
+// it removes the bad practice of using `init()`.
+
 package expvar